Refactor client.
